pyproject.toml
[project]
name = "mu-bench"
version = "0.1.0"
description = "Multilingual Utterances Transcription Benchmark"
requires-python = ">=3.11"
dependencies = ["pyyaml>=6.0"]

[project.optional-dependencies]
tools = ["huggingface-hub>=0.20", "requests>=2.28"]

scoring = [
    "jiwer>=3.0",
    "openai>=1.0",
    "requests>=2.28",
    "python-dotenv>=1.0",
    "pyyaml>=6.0",
]

transcribe = [
    "aiohttp>=3.9",
    "soundfile>=0.12",
    "numpy>=1.24",
    "openai>=1.0",
    "google-cloud-speech>=2.0",
    "huggingface-hub>=0.20",
    "python-dotenv>=1.0",
    "pyyaml>=6.0",
]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["scoring"]
